% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_js.R
\name{actionButton_toggle}
\alias{actionButton_toggle}
\title{Add a "Toggle" style to an actionButton}
\usage{
actionButton_toggle(inputId)
}
\description{
Call this button on the \code{inputId} of the action button with a reactive observer
}
\examples{
library(shiny)
devtools::load_all(pkgload::pkg_path())
ui <- fluidPage(
  shinyVirga::use_shinyVirga(),
  shiny::actionButton(
    'modal',
    "Open Modal",
    icon = shiny::icon("redo")
  )
)


# Define server logic required to draw a histogram
server <- function(input, output) {
  observeEvent(input$modal,
               {
                 actionButton_close("modal")
                 shiny::showModal(modalDialog(
                   title = "Hi, Modal",
                   footer = actionButton('close_modal',
                                         "Close")
                 ))
               })
  observeEvent(input$close_modal, {
    actionButton_close("modal")
    removeModal()
  })

}
shinyApp(ui = ui, server = server)
}
