% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_js.R
\name{actionButton_toggle_style}
\alias{actionButton_toggle_style}
\title{Add a "Toggle" style to an actionButton}
\usage{
actionButton_toggle_style(btn)
}
\arguments{
\item{btn}{\code{shiny.tag} The actionButton to style}
}
\value{
\code{shiny.tag.list} With the button
}
\description{
Once clicked, the button will appear to be in a "pressed" state until the next click.
}
