% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{add_module}
\alias{add_module}
\title{Add a module file to the \code{path} specified}
\usage{
add_module(
  name,
  path = "R",
  export = FALSE,
  m_name = sprintf("mod_\%s", name),
  m_ui = sprintf("mod_\%s_ui", name),
  m_server = sprintf("mod_\%s_server", name),
  use_ud = TRUE,
  ...
)
}
\arguments{
\item{m_name}{\code{chr} The name of file}

\item{m_ui}{\code{chr} The name of the module ui function}

\item{m_server}{\code{chr} The name of the module server function}
}
\description{
Add a module file to the \code{path} specified
}
