% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_css.R
\name{add_sass}
\alias{add_sass}
\title{Translate inline Sass to CSS style tag}
\usage{
add_sass(sass, .open = "|", .close = "|", .envir = rlang::caller_env())
}
\arguments{
\item{sass}{\code{chr} with Sass}

\item{.open}{[\code{character(1)}: \sQuote{\\{}]\cr The opening delimiter. Doubling the
full delimiter escapes it.}

\item{.close}{[\code{character(1)}: \sQuote{\\}}]\cr The closing delimiter. Doubling the
full delimiter escapes it.}

\item{.envir}{[\code{environment}: \code{parent.frame()}]\cr Environment to evaluate each expression in. Expressions are
evaluated from left to right. If \code{.x} is an environment, the expressions are
evaluated in that environment and \code{.envir} is ignored. If \code{NULL} is passed, it is equivalent to \code{\link[=emptyenv]{emptyenv()}}.}
}
\value{
\code{shiny-tag}
}
\description{
Translate inline Sass to CSS style tag
}
\examples{
ns <- function(x) paste0("blah-ble-",x)
add_sass("#|ns('myID')| > button {
  height: 10px;
  & option {
    text-align: center;
  }
}")
}
\seealso{
Other ui: 
\code{\link{acc_list}()},
\code{\link{box_list}()},
\code{\link{bs4Alert}()},
\code{\link{col_auto}()},
\code{\link{css_props}()},
\code{\link{dynamic_row}()},
\code{\link{enurl}()},
\code{\link{jq_hide}()},
\code{\link{list_to_li}()},
\code{\link{list_to_p}()},
\code{\link{make_action_button}()},
\code{\link{make_columns}()},
\code{\link{rep_br}()},
\code{\link{tagRemoveAttributes}()},
\code{\link{ui_row}()},
\code{\link{ui_tabs}()},
\code{\link{undisplay}()},
\code{\link{with_red_star}()}

Other css: 
\code{\link{css_props}()}
}
\concept{css}
\concept{ui}
