% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{browser_server}
\alias{browser_server}
\title{Add the \link[shiny]{observeEvent} call for a hidden ui browser button.}
\usage{
browser_server(e = rlang::caller_env())
}
\arguments{
\item{e}{The environment from which this function is called. \strong{Default} \link[rlang]{caller_env} which typically works as is.}
}
\description{
Add the \link[shiny]{observeEvent} call for a hidden ui browser button.
}
\seealso{
browser_ui, golem::browser_button

Other general: 
\code{\link{browser_ui}()},
\code{\link{is_shiny.tag}()},
\code{\link{ns_find}()},
\code{\link{path_strip_shiny}()},
\code{\link{path_strip_to}()},
\code{\link{tab_ns_extract}()},
\code{\link{warn_id}()}
}
\concept{general}
