% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bootstrap.R
\name{bs_extract_status}
\alias{bs_extract_status}
\title{Extract the bootstrap status of a Bootstrap tag that uses status}
\usage{
bs_extract_status(x)
}
\arguments{
\item{x}{\code{shiny.tag}}
}
\value{
\code{chr}
}
\description{
Extract the bootstrap status of a Bootstrap tag that uses status
}
