% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bootstrap.R
\docType{data}
\name{bs_statuses}
\alias{bs_statuses}
\title{Valid Bootstrap Statuses}
\format{
An object of class \code{character} of length 8.
}
\usage{
bs_statuses
}
\description{
Valid Bootstrap Statuses
}
\keyword{datasets}
