% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{card_grid}
\alias{card_grid}
\title{Create a gridded card layout}
\usage{
card_grid(..., icons = FALSE, top_row = NULL, description, infoboxes, graphs)
}
\arguments{
\item{...}{Any number of \link[shiny]{tag}s or \link[shiny]{tagList}s or \code{list}s (each treated as a separate entity to be housed in a row.). \code{shiny.tag}s and \code{shiny.tag.list}s, if named, will be inside of a \link[bs4Dash]{bs4Card} with the name as the title on it's own row. A \code{list} with named objects will be on a row with individual \link[bs4Dash]{bs4Card}s.}

\item{top_row}{\code{shiny.tag/shiny.tag.list} All elements belonging to the top row}

\item{description}{A shiny tag or tagList to be added to the description box}

\item{infoboxes}{\code{(named list/tagList)} with length divisible by 12. Will be added to the right of the description
\itemize{
\item{\code{named_list}}{ A named list of values to be put into infoboxes. Infobox will be titled by the name of the list item.}
\item{\code{tag/tagList}}{ shiny tag or tagList of items to include. These will individually be added to appropriate sized columns.}
}}

\item{graphs}{\code{shiny.tag/shiny.tag.list} Graphics to be placed in the further right column of the top summary area}
}
\value{
A \code{shiny::fluidPage} with the grid layout
}
\description{
Creates a grid layout informed by the \href{https://miro.com/app/board/o9J_l90QVck=/?moveToWidget=3074457361986467451&cot=14}{design spec}
}
