% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{col_auto}
\alias{col_auto}
\title{Render an auto-sized column}
\usage{
col_auto(..., .noWS = NULL, .renderHook = NULL)
}
\arguments{
\item{...}{Tag attributes (named arguments) and children (unnamed arguments).
A named argument with an \code{NA} value is rendered as a boolean attributes
(see example). Children may include any combination of:
\itemize{
\item Other tags objects
\item \code{\link[htmltools:HTML]{HTML()}} strings
\item \code{\link[htmltools:htmlDependency]{htmlDependency()}}s
\item Single-element atomic vectors
\item \code{list()}s containing any combination of the above
}}

\item{.noWS}{Character vector used to omit some of the whitespace that would
normally be written around this tag. Valid options include \code{before},
\code{after}, \code{outside}, \code{after-begin}, and \code{before-end}.
Any number of these options can be specified.}

\item{.renderHook}{A function (or list of functions) to call when the \code{tag} is rendered. This
function should have at least one argument (the \code{tag}) and return anything
that can be converted into tags via \code{\link[htmltools:as.tags]{as.tags()}}. Additional hooks may also be
added to a particular \code{tag} via \code{\link[htmltools:tagAddRenderHook]{tagAddRenderHook()}}.}
}
\value{
\code{shiny.tag}
}
\description{
Render an auto-sized column
}
\examples{
col_auto()
}
\seealso{
Other ui: 
\code{\link{acc_list}()},
\code{\link{add_sass}()},
\code{\link{box_list}()},
\code{\link{bs4Alert}()},
\code{\link{bs_extract_status}()},
\code{\link{bs_statuses}},
\code{\link{col_10}()},
\code{\link{col_11}()},
\code{\link{col_12}()},
\code{\link{col_1}()},
\code{\link{col_2}()},
\code{\link{col_3}()},
\code{\link{col_4}()},
\code{\link{col_5}()},
\code{\link{col_6}()},
\code{\link{col_7}()},
\code{\link{col_8}()},
\code{\link{col_9}()},
\code{\link{copyright}()},
\code{\link{css_props}()},
\code{\link{dynamic_row}()},
\code{\link{enurl}()},
\code{\link{fa_arrow_icon}()},
\code{\link{icon_sb}()},
\code{\link{infoIcon}()},
\code{\link{jq_hide}()},
\code{\link{list_to_li}()},
\code{\link{list_to_p}()},
\code{\link{make_action_button}()},
\code{\link{make_columns}()},
\code{\link{rep_br}()},
\code{\link{tagRemoveAttributes}()},
\code{\link{ui_row}()},
\code{\link{ui_tabs}()},
\code{\link{undisplay}()},
\code{\link{with_red_star}()}
}
\concept{ui}
