% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{copyright}
\alias{copyright}
\title{Creates an always up to date italicized copyright element.}
\usage{
copyright(
  copyright_holder = a("Virga Labs ", href = "https://www.virgalabs.io/", target =
    "_blank")
)
}
\value{
\code{shiny.tag}
}
\description{
Creates an always up to date italicized copyright element.
}
