% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_css.R
\name{css_props}
\alias{css_props}
\title{R List to CSS declarations as a string}
\usage{
css_props(..., declarations = NULL, inline = FALSE)
}
\arguments{
\item{...}{\code{chr} Named css properties, IE characters in the from \code{property = value}}

\item{declarations}{\code{list} CSS Declarations as a named list ie  \code{list(property = value)}}

\item{inline}{\code{lgl} Should CSS Properties be in inline format?}
}
\value{
\code{chr} properties as CSS character with newlines
}
\description{
R List to CSS declarations as a string
}
\examples{
css_props(display = "none", declarations = list(margin = "0px"))
}
\seealso{
Other ui: 
\code{\link{acc_list}()},
\code{\link{add_sass}()},
\code{\link{box_list}()},
\code{\link{bs4Alert}()},
\code{\link{bs_extract_status}()},
\code{\link{bs_statuses}},
\code{\link{col_10}()},
\code{\link{col_11}()},
\code{\link{col_12}()},
\code{\link{col_1}()},
\code{\link{col_2}()},
\code{\link{col_3}()},
\code{\link{col_4}()},
\code{\link{col_5}()},
\code{\link{col_6}()},
\code{\link{col_7}()},
\code{\link{col_8}()},
\code{\link{col_9}()},
\code{\link{col_auto}()},
\code{\link{copyright}()},
\code{\link{dynamic_row}()},
\code{\link{enurl}()},
\code{\link{fa_arrow_icon}()},
\code{\link{icon_sb}()},
\code{\link{infoIcon}()},
\code{\link{jq_hide}()},
\code{\link{list_to_li}()},
\code{\link{list_to_p}()},
\code{\link{make_action_button}()},
\code{\link{make_columns}()},
\code{\link{rep_br}()},
\code{\link{tagRemoveAttributes}()},
\code{\link{ui_row}()},
\code{\link{ui_tabs}()},
\code{\link{undisplay}()},
\code{\link{with_red_star}()}

Other css: 
\code{\link{add_sass}()},
\code{\link{list2sass}()}
}
\concept{css}
\concept{ui}
