% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_deploy.R
\name{deploy_stage}
\alias{deploy_stage}
\title{Run DMDU Deployment as a background task}
\usage{
deploy_stage(
  deploy_path = "deploy",
  copy_r_environ = TRUE,
  copy_r_profile = TRUE,
  copy_renv_lock = TRUE,
  lockfile_path = "renv.lock",
  copy_dockerfile = TRUE,
  GITHUB_PAT = remotes:::github_pat(),
  docker_image_tag = paste0(pkgload::pkg_name(), ":latest")
)
}
\arguments{
\item{deploy_path}{\code{chr} directory in which to stage deployment files}

\item{copy_r_environ}{\code{lgl} Whether to include the \emph{.Renviron} file}

\item{copy_r_profile}{\code{lgl} Whether to include the \emph{.Rprofile} file}

\item{copy_renv_lock}{\code{lgl} Whether to include the \emph{renv.lock} file}

\item{copy_dockerfile}{\code{lgl} Whether to include the \emph{Dockerfile}}
}
\value{
\code{job} Background job that deploys the dmdu app (not working locally)
}
\description{
Run DMDU Deployment as a background task
}
