% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{dynamic_row}
\alias{dynamic_row}
\title{Create a row with columns that dynamically resize
Must use \code{div(class = 'column', ...)} or \code{col_auto()} for internal columns to function properly.}
\usage{
dynamic_row(...)
}
\arguments{
\item{...}{Elements to include within the page}
}
\value{
A UI definition that can be passed to the \link[shiny]{shinyUI} function.
}
\description{
Create a row with columns that dynamically resize
Must use \code{div(class = 'column', ...)} or \code{col_auto()} for internal columns to function properly.
}
\examples{
dynamic_row(col_auto(bs4Dash::box(width = 12)), col_auto(bs4Dash::box(width = 12)))
}
\seealso{
Other ui: 
\code{\link{acc_list}()},
\code{\link{add_sass}()},
\code{\link{box_list}()},
\code{\link{bs4Alert}()},
\code{\link{col_auto}()},
\code{\link{css_props}()},
\code{\link{enurl}()},
\code{\link{jq_hide}()},
\code{\link{list_to_li}()},
\code{\link{list_to_p}()},
\code{\link{make_action_button}()},
\code{\link{make_columns}()},
\code{\link{rep_br}()},
\code{\link{tagRemoveAttributes}()},
\code{\link{ui_row}()},
\code{\link{ui_tabs}()},
\code{\link{undisplay}()},
\code{\link{with_red_star}()}
}
\concept{ui}
