% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_icons.R
\name{fa_arrow_icon}
\alias{fa_arrow_icon}
\title{FontAwesome Arrow Icons}
\usage{
fa_arrow_icon(
  direction = c("up", "down", "left-right", "up-down", "up-down-left-right")
)
}
\arguments{
\item{direction}{\code{chr} One of \code{up, down, left-right, up-down, up-down-left-right}}
}
\value{
\code{shiny.tag}
}
\description{
FontAwesome Arrow Icons
}
\examples{
fa_arrow_icon("up-down")
}
