% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reactives.R
\name{get_userData}
\alias{get_userData}
\title{Retreive a value from the shiny session userData environment}
\usage{
get_userData(x, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{x}{\code{chr} The name of the object to retrieve}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}. Default is \code{getDefaultReactiveDomain()}.}
}
\value{
the object
}
\description{
Retreive a value from the shiny session userData environment
}
\examples{
\notrun{
 # Must run inside a shiny app
 session$userData$myData <- c(1,2,3)
 print(get_userData("myData"))
}
}
\seealso{
Other reactives: 
\code{\link{new_rv}()},
\code{\link{nonreactiveVal}()},
\code{\link{rv}()},
\code{\link{rv_index}()},
\code{\link{rv_modify}()},
\code{\link{rv_to_list}()}
}
\concept{reactives}
