% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{glossary_sync}
\alias{glossary_sync}
\title{Update the Virga Labs glossary}
\usage{
glossary_sync(
  ss = "163ArY3cL67Vp-gzqjKSw_4r2kl-pCqMRsKDrK_zgbM0",
  sheet = "Main",
  file = "R/glossary.R"
)
}
\arguments{
\item{ss}{Something that identifies a Google Sheet:
\itemize{
\item its file id as a string or \code{\link[googledrive:as_id]{drive_id}}
\item a URL from which we can recover the id
\item a one-row \code{\link[googledrive:dribble]{dribble}}, which is how googledrive
represents Drive files
\item an instance of \code{googlesheets4_spreadsheet}, which is what \code{\link[googlesheets4:gs4_get]{gs4_get()}}
returns
}

Processed through \code{\link[googlesheets4:as_sheets_id]{as_sheets_id()}}.}

\item{sheet}{Sheet to read, in the sense of "worksheet" or "tab". You can identify a sheet by name, with a string, or by position, with a number. Ignored if the sheet is specified via \code{range}. If neither argument specifies the sheet, defaults to the first visible sheet.}

\item{file}{either a character string naming a file or a
    \link[base]{connection}.  \code{""} indicates output to the console.}
}
\value{
\code{chr} file path invisibly
}
\description{
Update the Virga Labs glossary
}
