% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{grid_card}
\alias{grid_card}
\title{Wrap in a \link[bs4Dash]{bs4Card} with default options}
\usage{
grid_card(
  ...,
  title = NULL,
  id = NULL,
  class = NULL,
  tip_icon = NULL,
  width = 12,
  maximizable = TRUE,
  elevation = 2,
  more_info = NULL,
  collapsible = FALSE
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[bs4Dash:box]{bs4Dash::bs4Card}}
  \describe{
    \item{\code{footer}}{Optional footer text.}
    \item{\code{status}}{The status of the item. This determines the item's background
  color.  Valid statuses are defined as follows:
\itemize{
  \item \code{primary}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#007bff")}.
  \item \code{secondary}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#6c757d")}.
  \item \code{info}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#17a2b8")}.
  \item \code{success}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#28a745")}.
  \item \code{warning}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#ffc107")}.
  \item \code{danger}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#dc3545")}.
  \item \code{gray-dark}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#343a40")}.
  \item \code{gray}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#adb5bd")}.
  \item \code{white}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#fff")}.
  \item \code{indigo}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#6610f2")}.
  \item \code{lightblue}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#3c8dbc")}.
  \item \code{navy}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#001f3f")}.
  \item \code{purple}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#605ca8")}.
  \item \code{fuchsia}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#f012be")}.
  \item \code{pink}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#e83e8c")}.
  \item \code{maroon}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#d81b60")}.
  \item \code{orange}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#ff851b")}.
  \item \code{lime}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#01ff70")}.
  \item \code{teal}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#39cccc")}.
  \item \code{olive}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#3d9970")}.
}}
    \item{\code{solidHeader}}{Should the header be shown with a solid color background?}
    \item{\code{background}}{If NULL (the default), the background of the box will be
  white. Otherwise, a color string. Valid colors are listed in
  \link[bs4Dash]{validColors}. See below:
\itemize{
 \item \code{primary}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#007bff")}.
  \item \code{secondary}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#6c757d")}.
  \item \code{info}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#17a2b8")}.
  \item \code{success}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#28a745")}.
  \item \code{warning}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#ffc107")}.
  \item \code{danger}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#dc3545")}.
  \item \code{gray-dark}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#343a40")}.
  \item \code{gray}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#adb5bd")}.
  \item \code{white}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#fff")}.
  \item \code{indigo}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#6610f2")}.
  \item \code{lightblue}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#3c8dbc")}.
  \item \code{navy}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#001f3f")}.
  \item \code{purple}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#605ca8")}.
  \item \code{fuchsia}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#f012be")}.
  \item \code{pink}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#e83e8c")}.
  \item \code{maroon}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#d81b60")}.
  \item \code{orange}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#ff851b")}.
  \item \code{lime}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#01ff70")}.
  \item \code{teal}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#39cccc")}.
  \item \code{olive}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#3d9970")}.
}}
    \item{\code{height}}{The height of a box, in pixels or other CSS unit. By default
the height scales automatically with the content.}
    \item{\code{collapsed}}{If TRUE, start collapsed. This must be used with
\code{collapsible=TRUE}.}
    \item{\code{closable}}{If TRUE, display a button in the upper right that allows the user to close the box.}
    \item{\code{icon}}{Header icon. Displayed before title. Expect \code{\link[shiny]{icon}}.}
    \item{\code{gradient}}{Whether to allow gradient effect for the background color. Default to FALSE.}
    \item{\code{boxToolSize}}{Size of the toolbox: choose among "xs", "sm", "md", "lg".}
    \item{\code{headerBorder}}{Whether to display a border between the header and body.
TRUE by default}
    \item{\code{label}}{Slot for \link[bs4Dash]{boxLabel}.}
    \item{\code{dropdownMenu}}{List of items in the boxtool dropdown menu. Use \link[bs4Dash]{boxDropdown}.}
    \item{\code{sidebar}}{Slot for \link[bs4Dash]{boxSidebar}.}
  }}

\item{title}{Optional title.}

\item{id}{Card id.}

\item{class}{A class to applied to the top level div}

\item{tip_icon}{Tip icon. Tooltip Icon displayed after title. Expect \code{\link[tippy]{tippy}}.}

\item{width}{The width of the box, using the Bootstrap grid system. This is
used for row-based layouts. The overall width of a region is 12, so the
default card width of 6 occupies 1/2 of that width. For column-based
layouts, use \code{NULL} for the width; the width is set by the column that
contains the box.}

\item{maximizable}{If TRUE, the card can be displayed in full screen mode.}

\item{elevation}{Card elevation.}

\item{collapsible}{If TRUE, display a button in the upper right that allows
the user to collapse the box.}
}
\description{
Wraps in a width 12 card, converts the title to an ID by replacing spaces with \code{_}. Adds a \code{tip_icon} if a tip named with the title exists in \code{tips}.
}
