% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{has_card}
\alias{has_card}
\title{Is a shiny.tag or shiny.tag.list have a Bootstrap card in the nesting structure?}
\usage{
has_card(x)
}
\arguments{
\item{x}{\code{shiny.tag/shiny.tag.list}}
}
\value{
\code{lgl}
}
\description{
Is a shiny.tag or shiny.tag.list have a Bootstrap card in the nesting structure?
}
\examples{
has_card(shiny::tagList(shiny::div(bs4Dash::bs4Card())))
}
