% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_icons.R
\name{icon_sb}
\alias{icon_sb}
\title{Wrapper for \link[shiny]{icon} that supports svg images}
\usage{
icon_sb(
  name,
  class = NULL,
  lib = "font-awesome",
  inactive_path = NULL,
  active_path = NULL,
  inactive_props = NULL,
  active_props = inactive_props,
  ...
)
}
\arguments{
\item{name}{\code{chr} FontAwesome name or image path}

\item{class}{Additional classes to customize the style of an icon (see the
\href{https://fontawesome.com/how-to-use}{usage examples} for details on
supported styles).}

\item{lib}{The icon library to use. Either \code{"font-awesome"} or \code{"glyphicon"}.}

\item{inactive_path}{\code{chr} path to the icon image file in it's inactive (unselected) state, starting with \code{www}.}

\item{active_path}{\code{chr} path to the icon image file in it's active (selected) state, starting with \code{www}.}

\item{...}{Arguments passed to the \verb{<i>} tag of \link[htmltools:builder]{htmltools::tags}.}
}
\value{
An \verb{<i>} (icon) HTML tag.
}
\description{
Wrapper for \link[shiny]{icon} that supports svg images
}
\seealso{
Other ui: 
\code{\link{acc_list}()},
\code{\link{add_sass}()},
\code{\link{box_list}()},
\code{\link{bs4Alert}()},
\code{\link{bs_extract_status}()},
\code{\link{bs_statuses}},
\code{\link{col_10}()},
\code{\link{col_11}()},
\code{\link{col_12}()},
\code{\link{col_1}()},
\code{\link{col_2}()},
\code{\link{col_3}()},
\code{\link{col_4}()},
\code{\link{col_5}()},
\code{\link{col_6}()},
\code{\link{col_7}()},
\code{\link{col_8}()},
\code{\link{col_9}()},
\code{\link{col_auto}()},
\code{\link{copyright}()},
\code{\link{css_props}()},
\code{\link{dynamic_row}()},
\code{\link{enurl}()},
\code{\link{fa_arrow_icon}()},
\code{\link{infoIcon}()},
\code{\link{jq_hide}()},
\code{\link{list_to_li}()},
\code{\link{list_to_p}()},
\code{\link{make_action_button}()},
\code{\link{make_columns}()},
\code{\link{rep_br}()},
\code{\link{tagRemoveAttributes}()},
\code{\link{ui_row}()},
\code{\link{ui_tabs}()},
\code{\link{undisplay}()},
\code{\link{with_red_star}()}
}
\concept{ui}
