% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_icons.R
\name{infoIcon}
\alias{infoIcon}
\title{A small info icon with a tooltip}
\usage{
infoIcon(..., name = "info-circle", class = NULL, lib = "font-awesome")
}
\arguments{
\item{...}{\code{shiny.tag.list/shiny.tag}s to be rendered as the tooltip}

\item{name}{The name of the icon. A name from either \href{https://fontawesome.com/}{Font Awesome} (when \code{lib="font-awesome"}) or
\href{https://getbootstrap.com/docs/3.3/components/#glyphicons}{Bootstrap Glyphicons} (when
\code{lib="glyphicon"}) may be provided. Note that the \code{"fa-"} and
\code{"glyphicon-"} prefixes should not appear in name (i.e., the
\code{"fa-calendar"} icon should be referred to as \code{"calendar"}). A \code{name} of
\code{NULL} may also be provided to get a raw \verb{<i>} tag with no library attached
to it.}

\item{class}{Additional classes to customize the style of an icon (see the
\href{https://fontawesome.com/how-to-use}{usage examples} for details on
supported styles).}

\item{lib}{The icon library to use. Either \code{"font-awesome"} or \code{"glyphicon"}.}
}
\value{
\code{shiny.tag} tippy enabled info icon
}
\description{
Requires \link[tippy]{use_tippy} to be placed in the \code{head} of the app. See the [tippy docs]{https://github.com/JohnCoene/tippy} for details
}
\examples{
infoIcon(tags$p("Help text here"))
}
\seealso{
Other ui: 
\code{\link{acc_list}()},
\code{\link{add_sass}()},
\code{\link{box_list}()},
\code{\link{bs4Alert}()},
\code{\link{bs_extract_status}()},
\code{\link{bs_statuses}},
\code{\link{col_10}()},
\code{\link{col_11}()},
\code{\link{col_12}()},
\code{\link{col_1}()},
\code{\link{col_2}()},
\code{\link{col_3}()},
\code{\link{col_4}()},
\code{\link{col_5}()},
\code{\link{col_6}()},
\code{\link{col_7}()},
\code{\link{col_8}()},
\code{\link{col_9}()},
\code{\link{col_auto}()},
\code{\link{copyright}()},
\code{\link{css_props}()},
\code{\link{dynamic_row}()},
\code{\link{enurl}()},
\code{\link{fa_arrow_icon}()},
\code{\link{icon_sb}()},
\code{\link{jq_hide}()},
\code{\link{list_to_li}()},
\code{\link{list_to_p}()},
\code{\link{make_action_button}()},
\code{\link{make_columns}()},
\code{\link{rep_br}()},
\code{\link{tagRemoveAttributes}()},
\code{\link{ui_row}()},
\code{\link{ui_tabs}()},
\code{\link{undisplay}()},
\code{\link{with_red_star}()}
}
\concept{ui}
