% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_golem.R
\name{is_shiny.tag.list}
\alias{is_shiny.tag.list}
\title{Is object a shiny \link[shiny]{tagList}}
\usage{
is_shiny.tag.list(x)
}
\arguments{
\item{x}{\code{object}}
}
\value{
\code{lgl} as to whether it's a \code{shiny.tag.list}
}
\description{
Is object a shiny \link[shiny]{tagList}
}
\examples{
is_shiny.tag.list(htmltools::tags$p())
is_shiny.tag.list(tagList(shiny::actionButton('a', 'a'), htmltools::tags$p()))
}
