% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_golem.R
\name{is_shiny.tag}
\alias{is_shiny.tag}
\title{Is object a shiny \link[shiny]{tag} or \link[shiny]{tagList}}
\usage{
is_shiny.tag(x)
}
\arguments{
\item{x}{\code{object}}
}
\value{
\code{(logical)}
}
\description{
Is object a shiny \link[shiny]{tag} or \link[shiny]{tagList}
}
\seealso{
Other golem: 
\code{\link{enurl}()},
\code{\link{jq_hide}()},
\code{\link{list_to_li}()},
\code{\link{list_to_p}()},
\code{\link{make_action_button}()},
\code{\link{rep_br}()},
\code{\link{tagRemoveAttributes}()},
\code{\link{undisplay}()},
\code{\link{with_red_star}()}

Other general: 
\code{\link{browser_server}()},
\code{\link{browser_ui}()},
\code{\link{ns_custom}()},
\code{\link{ns_find}()},
\code{\link{tab_ns_extract}()},
\code{\link{warn_id}()}
}
\concept{general}
\concept{golem}
