% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_js.R
\name{js_accordion_toggle}
\alias{js_accordion_toggle}
\title{Opens all accordions of provided \code{id}}
\usage{
js_accordion_toggle(id, asis = FALSE, .ns = ns_find())
}
\arguments{
\item{id}{\code{(chr)} Automatically namespaced with \code{ns} if inside a module}

\item{asis}{If \code{TRUE}, use the ID as-is even when inside a module
(instead of adding the namespace prefix to the ID).}

\item{.ns}{\code{fun} ns function. Typically found automatically.}
}
\value{
\code{none} Opens accordions
}
\description{
Opens all accordions of provided \code{id}
}
