% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_js.R
\name{js_all_plotly_visible}
\alias{js_all_plotly_visible}
\title{Check to ensure all plotlys are visible within an element}
\usage{
js_all_plotly_visible(id, asis = FALSE, .ns = ns_find())
}
\arguments{
\item{id}{\code{(chr)} Automatically namespaced with \code{ns} if inside a module}

\item{asis}{If \code{TRUE}, use the ID as-is even when inside a module
(instead of adding the namespace prefix to the ID).}

\item{.ns}{\code{fun} ns function. Typically found automatically.}
}
\value{
\code{none} called for side affects
}
\description{
Requires \code{shinyVirga::use_shinyVirga()} in the head of the DOM of the app (typically in ui file).
}
