% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_js.R
\name{js_bs4Card_action}
\alias{js_bs4Card_action}
\title{Open/Close a bs4Card with ID}
\usage{
js_bs4Card_action(id, action = "toggle")
}
\arguments{
\item{id}{\code{chr} ID of the bs4Card}

\item{action}{\code{chr} One of:
\itemize{
\item{\code{toggle}}{ Toggle Open/Closed}
\item{\code{open}}{ If closed, open, otherwise do nothing}
\item{\code{close}}{ If open, close, otherwise do nothing}
}}
}
\value{
\code{None}
}
\description{
Open/Close a bs4Card with ID
}
