% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_js.R
\name{js_callback}
\alias{js_callback}
\title{Create an anonymous JS function to monitor an event and bind it to a shiny input}
\usage{
js_callback(id, js = "", value, ..., asis = FALSE, .ns = ns_find())
}
\arguments{
\item{id}{\code{(chr)} Automatically namespaced with \code{ns} if inside a module}

\item{js}{\code{(chr)} path to js file with just the anonymous function, or js code to interpolate into a generalized anonymous javascript function that uses glue insertions demarcated by \code{{{}}}:
\preformatted{
(e) => {
         console.log('{{input}} event');
         {{js}}
Shiny.setInputValue('{{input}}', {{value}}, {priority: 'event'});
};}}

\item{value}{\code{(js)} arbitrary js code that defines the value to bind. Can also be a variable name if declared in the \code{js} argument}

\item{...}{\code{(additional named arguments)} corresponding to any arguments that need be interpolated with \link[glue]{glue}}

\item{asis}{If \code{TRUE}, use the ID as-is even when inside a module
(instead of adding the namespace prefix to the ID).}

\item{.ns}{\code{fun} ns function. Typically found automatically}
}
\value{
\code{(chr)} with javascript interpolated with \link[glue]{glue} arguments
}
\description{
Create an anonymous JS function to monitor an event and bind it to a shiny input
}
\examples{
js_callback("widget", "var myValue = e.someProperty;", "myValue")
}
\seealso{
Other JS: 
\code{\link{js_after}()},
\code{\link{js_mouseover_once}()},
\code{\link{js_set_input_val}()}
}
\concept{JS}
