% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_js.R
\name{js_callout}
\alias{js_callout}
\title{Create a driver.js callout
Must include \code{shinyVirga::use_driver.js()} in the head of the page.}
\usage{
js_callout(
  id,
  title,
  description,
  position = "bottom-center",
  className = "scoped-class",
  animate = TRUE,
  opacity = 0.75,
  padding = 10,
  allowClose = TRUE,
  overlayClickNext = FALSE,
  doneBtnText = "Done",
  closeBtnText = "Close",
  stageBackground = "#ffffff",
  nextBtnText = "Next",
  prevBtnText = "Previous",
  showButtons = TRUE,
  keyboardControl = TRUE,
  scrollIntoViewOptions = list(),
  onHighlightStarted = NULL,
  onHighlighted = NULL,
  onDeselected = NULL,
  onReset = NULL,
  onNext = NULL,
  onPrevious = NULL,
  asis = FALSE,
  .ns = ns_find()
)
}
\arguments{
\item{id}{\code{(chr)} Automatically namespaced with \code{ns} if inside a module}

\item{className}{className to wrap driver.js popover}

\item{animate}{Whether to animate or not}

\item{opacity}{Background opacity (0 means only popovers and without overlay)}

\item{padding}{Distance of element from around the edges}

\item{allowClose}{Whether the click on overlay should close or not}

\item{overlayClickNext}{Whether the click on overlay should move next}

\item{doneBtnText}{Text on the final button}

\item{closeBtnText}{Text on the close button for this step}

\item{stageBackground}{Background color for the staged behind highlighted element}

\item{nextBtnText}{Next button text for this step}

\item{prevBtnText}{Previous button text for this step}

\item{showButtons}{Do not show control buttons in footer}

\item{keyboardControl}{Allow controlling through keyboard (escape to close, arrow keys to move)}

\item{scrollIntoViewOptions}{We use \code{scrollIntoView()} when possible, pass here the options for it if you want any}

\item{onHighlightStarted}{Called when element is about to be highlighted}

\item{onHighlighted}{Called when element is fully highlighted}

\item{onDeselected}{Called when element has been deselected}

\item{onReset}{Called when overlay is about to be cleared}

\item{onNext}{Called when moving to next step on any step}

\item{onPrevious}{Called when moving to previous step on any step}

\item{asis}{If \code{TRUE}, use the ID as-is even when inside a module
(instead of adding the namespace prefix to the ID).}

\item{.ns}{\code{fun} ns function. Typically found automatically if inside a \code{server} module. \strong{It is recommended to supply this argument if using this function in a \code{ui} module.}}
}
\description{
Create a driver.js callout
Must include \code{shinyVirga::use_driver.js()} in the head of the page.
}
