% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_js.R
\name{js_glow}
\alias{js_glow}
\title{Add an animated glow to an element}
\usage{
js_glow(id, color = "deepskyblue", asis = FALSE, delay = NULL)
}
\arguments{
\item{id}{\code{(chr)} Automatically namespaced with \code{ns} if inside a module}

\item{color}{\code{chr} A hexadecimal color to animate with}

\item{asis}{If \code{TRUE}, use the ID as-is even when inside a module
(instead of adding the namespace prefix to the ID).}

\item{delay}{\code{dbl} duration in ms before removing the content from the element. Set to 0 to allow the element to persist.}
}
\description{
Add an animated glow to an element
}
