% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_js.R
\name{js_mouseover_once}
\alias{js_mouseover_once}
\title{Make a shiny input with named ID with a logical TRUE value when the element is moused over.}
\usage{
js_mouseover_once(id, asis = FALSE, .ns = ns_find())
}
\arguments{
\item{id}{\code{(chr)} Automatically namespaced with \code{ns} if inside a module}

\item{asis}{If \code{TRUE}, use the ID as-is even when inside a module
(instead of adding the namespace prefix to the ID).}

\item{.ns}{\code{fun} ns function. Typically found automatically if inside a \code{server} module. \strong{It is recommended to supply this argument if using this function in a \code{ui} module.}}
}
\description{
Make a shiny input with named ID with a logical TRUE value when the element is moused over.
}
\seealso{
Other JS: 
\code{\link{js_after}()},
\code{\link{js_callback}()},
\code{\link{js_callout}()},
\code{\link{js_glow}()},
\code{\link{js_set_input_val}()},
\code{\link{use_driver.js}()}
}
\concept{JS}
