% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_js.R
\name{js_picker_disable}
\alias{js_picker_disable}
\title{Disable a bootstrap pickerInput}
\usage{
js_picker_disable(id, asis = FALSE, .ns = ns_find())
}
\arguments{
\item{id}{\code{(chr)} Automatically namespaced with \code{ns} if inside a module}

\item{asis}{If \code{TRUE}, use the ID as-is even when inside a module
(instead of adding the namespace prefix to the ID).}

\item{.ns}{\code{fun} ns function. Typically found automatically.}
}
\value{
\code{none} called for side-effects
}
\description{
Disable a bootstrap pickerInput
}
\seealso{
Other JS: 
\code{\link{js_after}()},
\code{\link{js_callback}()},
\code{\link{js_callout}()},
\code{\link{js_glow}()},
\code{\link{js_mouseover_once}()},
\code{\link{js_picker_enable}()},
\code{\link{js_set_input_val}()},
\code{\link{use_driver.js}()}
}
\author{
Maria Sevillano, Stephen Holsenbeck
}
\concept{JS}
