% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_js.R
\name{js_set_attrs}
\alias{js_set_attrs}
\title{Set attributes of a DOM element}
\usage{
js_set_attrs(id, ..., asis = FALSE, .ns = ns_find())
}
\arguments{
\item{id}{\code{(chr)} Automatically namespaced with \code{ns} if inside a module}

\item{...}{\code{named args} attributes to set}

\item{asis}{If \code{TRUE}, use the ID as-is even when inside a module
(instead of adding the namespace prefix to the ID).}

\item{.ns}{\code{fun} ns function. Typically found automatically.}
}
\value{
\code{None} Updates the element
}
\description{
This will replace existing attribute if it exists.
}
