% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_js.R
\name{js_sliderInput_ticks}
\alias{js_sliderInput_ticks}
\title{Remove or show sliderInput tick marks}
\usage{
js_sliderInput_ticks(id, hide = TRUE, asis = FALSE, .ns = ns_find())
}
\arguments{
\item{id}{\code{(chr)} Automatically namespaced with \code{ns} if inside a module}

\item{hide}{\code{lgl} Whether to hide the tick marks}

\item{asis}{If \code{TRUE}, use the ID as-is even when inside a module
(instead of adding the namespace prefix to the ID).}

\item{.ns}{\code{fun} ns function. Typically found automatically.}
}
\value{
\code{None} Removes tick marks from slider
}
\description{
Remove or show sliderInput tick marks
}
