% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_css.R
\name{list2sass}
\alias{list2sass}
\title{Conver a list or vector to Sass variables}
\usage{
list2sass(x, pre = "", suf = "")
}
\arguments{
\item{x}{\code{list/chr}}

\item{pre}{\code{chr} prefix to prepend}

\item{suf}{\code{chr} suffix to append}
}
\value{
\code{chr}
}
\description{
Conver a list or vector to Sass variables
}
\examples{
list2sass(pre = "dark-", list(
  lightblue = "rgba(18,180,211,1)",
  darkblue = "rgba(2,120,170,1)",
  navyblue = "rgba(0,57,73,1)",
  brown = "rgba(72,36,18,1)",
  lightbrown = "rgb(132,96,78)",
  purple = "rgba(111,96,140,1)",
  lightpurple = "rgba(165,150,194,1)",
  darkcyan = "rgba(0,166,212,1)",
  darkturquoise = "rgba(9,119,168,1)"
)) |> cat(sep = '\n')
}
\seealso{
Other css: 
\code{\link{add_sass}()},
\code{\link{css_props}()}
}
\concept{css}
