% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{make_columns}
\alias{make_columns}
\title{Make columns from assorted shiny.tag elements
Sorts shiny.tags into columns based on the maximum number of columns (\code{max_cols}) per row}
\usage{
make_columns(x, max_cols = TRUE, fn = list(bs4Dash::box, bs4Dash::column)[[1]])
}
\arguments{
\item{x}{\code{(shiny.tags)}}

\item{max_cols}{\code{(logical/integer)} Either \code{TRUE} \strong{Default} for a default of 4 columns per row, \code{FALSE} for no columns, or an integer indicating the max number of columns.}

\item{fn}{\code{fun} with which to wrap each column's content. Typically \link[bs4Dash]{box} or \link[bs4Dash]{column}}
}
\value{
\code{(list(s))}
}
\description{
Make columns from assorted shiny.tag elements
Sorts shiny.tags into columns based on the maximum number of columns (\code{max_cols}) per row
}
\seealso{
Other ui: 
\code{\link{acc_list}()},
\code{\link{add_sass}()},
\code{\link{box_list}()},
\code{\link{bs4Alert}()},
\code{\link{col_auto}()},
\code{\link{css_props}()},
\code{\link{dynamic_row}()},
\code{\link{enurl}()},
\code{\link{jq_hide}()},
\code{\link{list_to_li}()},
\code{\link{list_to_p}()},
\code{\link{make_action_button}()},
\code{\link{rep_br}()},
\code{\link{tagRemoveAttributes}()},
\code{\link{ui_row}()},
\code{\link{ui_tabs}()},
\code{\link{undisplay}()},
\code{\link{with_red_star}()}
}
\concept{ui}
