% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_translation.R
\name{mod_transifex_select}
\alias{mod_transifex_select}
\title{Add Transifex Language selector}
\usage{
mod_transifex_select(
  inputId,
  label,
  choices = c(English = "en", Spanish = "es"),
  selected = c(English = "en"),
  multiple = FALSE,
  options = list(),
  choicesOpt = NULL,
  width = NULL,
  class = NULL,
  inline = FALSE,
  transifex_options = list(detect_lang = FALSE, picker = "#test", version = "latest"),
  key = Sys.getenv("TRANSIFEX_API_KEY", ""),
  ...
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{choices}{List of values to select from. If elements of the
list are named then that name rather than the value is displayed to the user.}

\item{selected}{The initially selected value (or multiple values if \code{multiple = TRUE}).
If not specified then defaults to the first value for single-select lists
and no values for multiple select lists.}

\item{multiple}{Is selection of multiple items allowed?}

\item{options}{List of options, see \link[shinyWidgets]{pickerOptions} for all available options.
To limit the number of selection possible, see example below.}

\item{choicesOpt}{Options for choices in the dropdown menu.}

\item{width}{The width of the input : 'auto', 'fit', '100px', '75\%'.}

\item{inline}{Display picker inline, to have label and menu on same line use \code{width = "fit"}.}

\item{transifex_options}{\code{list} List of options to be passed to the \href{https://help.transifex.com/en/articles/6370718-javascript-api}{liveSettings object}}

\item{...}{a collection of attributes.}
}
\value{
\code{shiny.tag}
}
\description{
Must set up an environment variable entitled \code{TRANSIFEX_API_KEY} with your API Key created \href{https://app.transifex.com/user/settings/api/}{here} and run \code{transifex_select_server} at least once in your app.
}
\details{
Implementation adapted from the \href{https://help.transifex.com/en/articles/6259014-technical-instructions#h_aa6dc907c4}{Transifex docs}
}
\seealso{
Other translation: 
\code{\link{mod_transifex_select_server}()}
}
\concept{translation}
