% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{msg_mod_fun}
\alias{msg_mod_fun}
\title{Display the name of the parent module}
\usage{
msg_mod_fun(.call = rlang::trace_back(bottom = 5), e = rlang::caller_env())
}
\value{
Message with the name of the running module and it's environment
}
\description{
Stack traces often aren't available in Shiny. Use this function inside of modules to know where errors occur
}
\seealso{
Other debugging: 
\code{\link{browser_server}()},
\code{\link{browser_ui}()},
\code{\link{use_msg_mod_fun}()}
}
\concept{debugging}
