% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reactives.R
\name{new_rv}
\alias{new_rv}
\title{Create a \code{\link[shiny]{reactiveVal}} or a \code{nonreactiveVal} depending on whether shiny is running}
\usage{
new_rv(val = NULL)
}
\arguments{
\item{val}{\code{obj} any initial value}
}
\value{
\code{reactiveVal/nonreactiveVal}
}
\description{
Create a \code{\link[shiny]{reactiveVal}} or a \code{nonreactiveVal} depending on whether shiny is running
}
\examples{
new_rv(0)
}
\seealso{
Other reactives: 
\code{\link{get_userData}()},
\code{\link{nonreactiveVal}()},
\code{\link{rv}()},
\code{\link{rv_index}()},
\code{\link{rv_modify}()},
\code{\link{rv_to_list}()}
}
\concept{reactives}
