% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{nm_to_id}
\alias{nm_to_id}
\title{Convert a string to a valid HTML ID}
\usage{
nm_to_id(x)
}
\arguments{
\item{x}{\code{chr}}
}
\value{
\code{chr}
}
\description{
Convert a string to a valid HTML ID
}
\examples{
nm_to_id("A random string")
}
