% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reactives.R
\name{nonreactiveVal}
\alias{nonreactiveVal}
\alias{nrV}
\title{Create a non-reactiveVal}
\usage{
nonreactiveVal(value = NULL)
}
\arguments{
\item{value}{An optional initial value.}
}
\value{
\code{obj} Whatever object was last stored
}
\description{
these are useful in conjunction with \code{req} to evaluate whether a reactive should run while preventing reactive flush or chaining when the internal value changes
}
\examples{
my_nrv <- nonreactiveVal(5)
my_nrv()
}
