% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{ns_custom}
\alias{ns_custom}
\title{Create a custom namespace string.}
\usage{
ns_custom(levels, add = NULL, .ns = ns_find())
}
\arguments{
\item{levels}{\code{num} indicating what segments of the namespace to keep.}

\item{add}{\code{chr} segments to append to the result}

\item{.ns}{\code{fun} ns function. Typically found automatically.}
}
\value{
\code{chr} The custom namespace string
}
\description{
Useful for referencing objects nested in adjacent modules
}
\details{
No \code{-} can be used in user supplied namespace names as this is the separator shiny uses to separate namespace sections and is how this function splits apart sections.
}
\examples{
.ns <- function(x) paste0("body-vulnerability-vuln_num_summary-vuln_summary_nums", "-", x)
ns_custom(-2, add = "blah", .ns = .ns)
}
\seealso{
Other general: 
\code{\link{browser_server}()},
\code{\link{browser_ui}()},
\code{\link{is_shiny.tag}()},
\code{\link{ns_find}()},
\code{\link{tab_ns_extract}()},
\code{\link{warn_id}()}
}
\concept{general}
