% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{ns_find}
\alias{ns_find}
\title{Retrieve the \code{ns} function}
\usage{
ns_find(e = rlang::caller_env())
}
\arguments{
\item{e}{\code{env}}
}
\value{
\code{fun}
}
\description{
Retrieve the \code{ns} function
}
\seealso{
Other general: 
\code{\link{browser_server}()},
\code{\link{browser_ui}()},
\code{\link{is_shiny.tag}()},
\code{\link{path_strip_shiny}()},
\code{\link{path_strip_to}()},
\code{\link{tab_ns_extract}()},
\code{\link{warn_id}()}
}
\concept{general}
