% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{row_wrap.shiny.tag.list}
\alias{row_wrap.shiny.tag.list}
\title{S3 Method for row_wrap shiny.tag.list}
\usage{
\method{row_wrap}{shiny.tag.list}(x, nm, icons)
}
\description{
S3 Method for row_wrap shiny.tag.list
}
