% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reactives.R
\name{rv_to_list}
\alias{rv_to_list}
\alias{rv2l}
\title{Convert a reactiveValues object to a list}
\usage{
rv_to_list(x, all.names = FALSE)

rv2l(x, all.names = FALSE)
}
\arguments{
\item{x}{A \code{reactivevalues} object.}

\item{all.names}{If \code{TRUE}, include objects with a leading dot. If \code{FALSE}
(the default) don't include those objects.}
}
\description{
Convenience wrapper around \link[shiny]{reactiveValuesToList}
}
\examples{
values <- reactiveValues(a = 1)
\dontrun{
reactiveValuesToList(values)
}

# To get the objects without taking dependencies on them, use isolate().
# isolate() can also be used when calling from outside a reactive context (e.g.
# at the console)
isolate(reactiveValuesToList(values))
}
\seealso{
Other reactives: 
\code{\link{rv_index}()},
\code{\link{rv}()}
}
\concept{reactives}
