% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_inputs.R
\name{select_choices}
\alias{select_choices}
\title{make a nested category list for \code{\link[shiny]{selectInput}}}
\usage{
select_choices(
  metrics_df,
  metric_col_values,
  category_col,
  metric_col_names,
  sort_within_categories = FALSE,
  sort_general = FALSE
)
}
\arguments{
\item{metrics_df}{\code{tbl} See description}

\item{metric_col_values}{\code{chr} column name with values}

\item{category_col}{\code{chr} column name with categories}

\item{metric_col_names}{\code{chr} column name with names (if there are names)}

\item{sort_within_categories}{\code{lgl} Whether to sort values within categories}

\item{sort_general}{\code{lgl} Whether to sort categories (if categories included), or the choices themselves if no categories.}
}
\value{
\code{chr/list} list if category_col supplied, otherwise character vector
}
\description{
Takes a "metrics-styled" \code{data.frame} with
}
\concept{inputs
\itemize{
\item{A column that indicates the category}
\item{A column with the values}
\item{A column with the names displayed to the user, if different from the values (\emph{optional})}
}
and returns a nested list appropriate for the \code{choices} argument to a \code{\link[shiny]{selectInput}}.}
