% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyVirga-package.R
\docType{package}
\name{shinyVirga-package}
\alias{shinyVirga}
\alias{shinyVirga-package}
\title{shinyVirga: A Collection of R shiny utilities by Virga Labs}
\description{
A collection of useful utility functions for working with shiny and shiny frameworks
}
\seealso{
Useful links:
\itemize{
  \item \url{https://martin-mccoy.github.io/shinyVirga/}
}

}
\author{
\strong{Maintainer}: Stephen Holsenbeck \email{stephen@virgalabs.io}

}
\keyword{internal}
