% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_js.R
\name{sidebar_toggle}
\alias{sidebar_toggle}
\title{Disable/enable the bs4dash sidebar}
\usage{
sidebar_toggle(sidebar_id = "sidebar", disable = TRUE)
}
\arguments{
\item{disable}{\code{lgl} Whether to disable the sidebar, FALSE to reenable}
}
\value{
\code{None}
}
\description{
Disable/enable the bs4dash sidebar
}
