% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{simpleCard}
\alias{simpleCard}
\title{Construct a simple Bootstrap card}
\usage{
simpleCard(..., width = 12, class = NULL, id = NULL)
}
\arguments{
\item{...}{\code{shiny.tag}s}

\item{width}{\code{num} Width of the encapsulating column}

\item{class}{\code{chr} any additional classes added to the div.card}

\item{id}{\code{chr} added to the div.card}
}
\value{
\code{shiny.tag}
}
\description{
Construct a simple Bootstrap card
}
