% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{tab_ns_extract}
\alias{tab_ns_extract}
\title{Extract the tabname to which this module instance corresponds from the ns}
\usage{
tab_ns_extract(
  extract,
  which = c("first", "last", "all")[1],
  ns_fun = ns_find(e),
  e = rlang::caller_env()
)
}
\arguments{
\item{ns_fun}{\code{fun} \code{ns} found within the module}
}
\value{
\code{chr} the namespace as string
}
\description{
Extract the tabname to which this module instance corresponds from the ns
}
\seealso{
Other general: 
\code{\link{browser_server}()},
\code{\link{browser_ui}()},
\code{\link{is_shiny.tag}()},
\code{\link{ns_custom}()},
\code{\link{ns_find}()},
\code{\link{path_strip_shiny}()},
\code{\link{path_strip_to}()},
\code{\link{warn_id}()}
}
\concept{general}
