% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_golem.R
\name{tagRemoveAttributes}
\alias{tagRemoveAttributes}
\title{Remove a tag attribute}
\usage{
tagRemoveAttributes(tag, ...)
}
\arguments{
\item{tag}{the tag}

\item{...}{the attributes to remove}
}
\value{
a new tag
}
\description{
Remove a tag attribute
}
\examples{
a <- shiny::tags$p(src = "plop", "pouet")
tagRemoveAttributes(a, "src")
}
\seealso{
Other golem: 
\code{\link{enurl}()},
\code{\link{is_shiny.tag}()},
\code{\link{jq_hide}()},
\code{\link{list_to_li}()},
\code{\link{list_to_p}()},
\code{\link{make_action_button}()},
\code{\link{rep_br}()},
\code{\link{undisplay}()},
\code{\link{with_red_star}()}

Other ui: 
\code{\link{acc_list}()},
\code{\link{add_sass}()},
\code{\link{box_list}()},
\code{\link{bs4Alert}()},
\code{\link{col_auto}()},
\code{\link{css_props}()},
\code{\link{dynamic_row}()},
\code{\link{enurl}()},
\code{\link{jq_hide}()},
\code{\link{list_to_li}()},
\code{\link{list_to_p}()},
\code{\link{make_action_button}()},
\code{\link{make_columns}()},
\code{\link{rep_br}()},
\code{\link{ui_row}()},
\code{\link{ui_tabs}()},
\code{\link{undisplay}()},
\code{\link{with_red_star}()}
}
\concept{golem}
\concept{ui}
