% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_translation.R
\name{transifex_select_server}
\alias{transifex_select_server}
\title{Add Transifex API Key to Javascript environment}
\usage{
transifex_select_server(
  inputId,
  key = Sys.getenv("TRANSIFEX_API_KEY", ""),
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{key}{\code{chr} Your Transifex API Key. Will be derived from an environment variable entitled \code{TRANSIFEX_API_KEY} if not specified}
}
\value{
\code{shiny.tag}
}
\description{
Must set up an environment variable entitled \code{TRANSIFEX_API_KEY} with your API Key created \href{https://app.transifex.com/user/settings/api/}{here} and run \code{transifex_select_server} at least once in your app. \strong{Requires shinyjs}
}
\seealso{
Other translation: 
\code{\link{transifex_select}()}
}
\concept{translation}
