% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_translation.R
\name{transifex_update_language}
\alias{transifex_update_language}
\title{Update the language}
\usage{
transifex_update_language(
  inputId,
  language,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{language}{\code{chr} Language code, usually two letters}
}
\value{
\code{none}
}
\description{
Updates the language picker and the Transifex dependencies running in the app.
}
