% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{ui_tabs}
\alias{ui_tabs}
\title{Create a \code{shiny.tag.list} of \link[bs4Dash]{bs4SidebarMenuItem}s with an input \link[tibble]{tibble}}
\usage{
ui_tabs(tabs)
}
\arguments{
\item{tabs}{\code{tbl} similar to the following:
\preformatted{
tabs <- tibble::tribble(
  ~ text, ~ icon, ~ dots,
  "Welcome", "house", list(),
  "Education", "chalkboard", list(),
  "Performance", "flag-checkered", list(),
  "Robustness", "circle", list(),
  "Vulnerabilty", "bomb", list()
  ) |>
    dplyr::mutate(tabName = snakecase::to_snake_case(text))
}
where \code{dots} is a list of additional arguments to \link[bs4Dash]{bs4SidebarMenuItem}}
}
\value{
\code{shiny.tag.list}
}
\description{
Create a \code{shiny.tag.list} of \link[bs4Dash]{bs4SidebarMenuItem}s with an input \link[tibble]{tibble}
}
\seealso{
tabs

Other ui: 
\code{\link{acc_list}()},
\code{\link{add_sass}()},
\code{\link{box_list}()},
\code{\link{bs4Alert}()},
\code{\link{col_10}()},
\code{\link{col_11}()},
\code{\link{col_12}()},
\code{\link{col_1}()},
\code{\link{col_2}()},
\code{\link{col_3}()},
\code{\link{col_4}()},
\code{\link{col_5}()},
\code{\link{col_6}()},
\code{\link{col_7}()},
\code{\link{col_8}()},
\code{\link{col_9}()},
\code{\link{col_auto}()},
\code{\link{css_props}()},
\code{\link{dynamic_row}()},
\code{\link{enurl}()},
\code{\link{jq_hide}()},
\code{\link{list_to_li}()},
\code{\link{list_to_p}()},
\code{\link{make_action_button}()},
\code{\link{make_columns}()},
\code{\link{rep_br}()},
\code{\link{tagRemoveAttributes}()},
\code{\link{ui_row}()},
\code{\link{undisplay}()},
\code{\link{with_red_star}()}
}
\concept{ui}
