% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_golem.R
\name{undisplay}
\alias{undisplay}
\alias{display}
\title{Hide or display a tag}
\usage{
undisplay(tag)

display(tag)
}
\arguments{
\item{tag}{the tag}
}
\value{
a tag
}
\description{
Hide or display a tag
}
\examples{
## Hide
a <- shiny::tags$p(src = "plop", "pouet")
undisplay(a)
b <- shiny::actionButton("go_filter", "go")
undisplay(b)

}
\seealso{
Other golem: 
\code{\link{enurl}()},
\code{\link{is_shiny.tag}()},
\code{\link{jq_hide}()},
\code{\link{list_to_li}()},
\code{\link{list_to_p}()},
\code{\link{make_action_button}()},
\code{\link{rep_br}()},
\code{\link{tagRemoveAttributes}()},
\code{\link{with_red_star}()}

Other ui: 
\code{\link{acc_list}()},
\code{\link{add_sass}()},
\code{\link{box_list}()},
\code{\link{bs4Alert}()},
\code{\link{col_auto}()},
\code{\link{css_props}()},
\code{\link{dynamic_row}()},
\code{\link{enurl}()},
\code{\link{jq_hide}()},
\code{\link{list_to_li}()},
\code{\link{list_to_p}()},
\code{\link{make_action_button}()},
\code{\link{make_columns}()},
\code{\link{rep_br}()},
\code{\link{tagRemoveAttributes}()},
\code{\link{ui_row}()},
\code{\link{ui_tabs}()},
\code{\link{with_red_star}()}
}
\concept{golem}
\concept{ui}
