% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{use_msg_mod_fun}
\alias{use_msg_mod_fun}
\title{Insert module debugging statements throughout module files.}
\usage{
use_msg_mod_fun(pattern = "^mod")
}
\arguments{
\item{pattern}{\code{chr} The pattern to search for in files that will be modified.}
}
\value{
Modifies all files beginning with \code{pattern} insert \link[shinyVirga]{msg_mod_fun}
}
\description{
Insert module debugging statements throughout module files.
}
\seealso{
Other debugging: 
\code{\link{browser_server}()},
\code{\link{browser_ui}()},
\code{\link{msg_mod_fun}()}
}
\concept{debugging}
