% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyVirga-package.R
\name{use_shinyVirga}
\alias{use_shinyVirga}
\title{Adds JS Dependencies to the DOM for shinyVirga to function}
\usage{
use_shinyVirga()
}
\value{
\code{tagList}
}
\description{
Adds JS Dependencies to the DOM for shinyVirga to function
}
