% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyVirga-package.R
\name{useShinyVirga}
\alias{useShinyVirga}
\title{Adds JS Dependencies to the DOM for shinyVirga to function}
\usage{
useShinyVirga()
}
\value{
\code{tagList}
}
\description{
Adds JS Dependencies to the DOM for shinyVirga to function
}
