% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{virga_copyright}
\alias{virga_copyright}
\title{Creates an always up to date italicized Virga copyright element.}
\usage{
virga_copyright()
}
\value{
\code{shiny.tag}
}
\description{
Creates an always up to date italicized Virga copyright element.
}
