% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{warn_id}
\alias{warn_id}
\title{Display a warning in a DOM element of choice}
\usage{
warn_id(warn_html, id, ..., ms = 10000)
}
\arguments{
\item{id}{\code{chr} ID of the DOM element in which to display}

\item{...}{
  Arguments passed on to \code{\link[shinyjs:html]{shinyjs::html}}
  \describe{
    \item{\code{html}}{The HTML/text to place inside the element. Can be either simple
plain text or valid HTML code.}
    \item{\code{add}}{If \code{TRUE}, then append \code{html} to the contents of the element;
otherwise overwrite it.}
    \item{\code{selector}}{JQuery selector of the elements to target. Ignored if the \code{id}
argument is given.}
    \item{\code{asis}}{If \code{TRUE}, use the ID as-is even when inside a module
(instead of adding the namespace prefix to the ID).}
  }}

\item{ms}{The number of milliseconds to wait (1000 milliseconds = 1 second)
before running the expression.}

\item{warn_text}{\code{chr/shiny.tags} HTML or shiny.tags to render to HTML to display in the DOM}
}
\value{
A warning on the specified DOM element
}
\description{
Display a warning in a DOM element of choice
}
\seealso{
Other general: 
\code{\link{browser_server}()},
\code{\link{browser_ui}()},
\code{\link{is_shiny.tag}()},
\code{\link{ns_custom}()},
\code{\link{ns_find}()},
\code{\link{tab_ns_extract}()}
}
\concept{general}
