% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_golem.R
\name{with_red_star}
\alias{with_red_star}
\title{Add a red star at the end of the text}
\usage{
with_red_star(text)
}
\arguments{
\item{text}{the HTLM text to put before the red star}
}
\value{
an html element
}
\description{
Adds a red star at the end of the text
(for example for indicating mandatory fields).
}
\examples{
with_red_star("Enter your name here")

}
\seealso{
Other golem: 
\code{\link{enurl}()},
\code{\link{is_shiny.tag}()},
\code{\link{jq_hide}()},
\code{\link{list_to_li}()},
\code{\link{list_to_p}()},
\code{\link{make_action_button}()},
\code{\link{rep_br}()},
\code{\link{tagRemoveAttributes}()},
\code{\link{undisplay}()}

Other ui: 
\code{\link{acc_list}()},
\code{\link{add_sass}()},
\code{\link{box_list}()},
\code{\link{bs4Alert}()},
\code{\link{bs_extract_status}()},
\code{\link{bs_statuses}},
\code{\link{col_10}()},
\code{\link{col_11}()},
\code{\link{col_12}()},
\code{\link{col_1}()},
\code{\link{col_2}()},
\code{\link{col_3}()},
\code{\link{col_4}()},
\code{\link{col_5}()},
\code{\link{col_6}()},
\code{\link{col_7}()},
\code{\link{col_8}()},
\code{\link{col_9}()},
\code{\link{col_auto}()},
\code{\link{copyright}()},
\code{\link{css_props}()},
\code{\link{dynamic_row}()},
\code{\link{enurl}()},
\code{\link{fa_arrow_icon}()},
\code{\link{icon_sb}()},
\code{\link{infoIcon}()},
\code{\link{jq_hide}()},
\code{\link{list_to_li}()},
\code{\link{list_to_p}()},
\code{\link{make_action_button}()},
\code{\link{make_columns}()},
\code{\link{rep_br}()},
\code{\link{tagRemoveAttributes}()},
\code{\link{ui_row}()},
\code{\link{ui_tabs}()},
\code{\link{undisplay}()}
}
\concept{golem}
\concept{ui}
