% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{write_col_fns}
\alias{write_col_fns}
\alias{col_1}
\alias{col_2}
\alias{col_3}
\alias{col_4}
\alias{col_5}
\alias{col_6}
\alias{col_7}
\alias{col_8}
\alias{col_9}
\alias{col_10}
\alias{col_11}
\alias{col_12}
\title{Write wrappers}
\usage{
write_col_fns(file = "R/utils_col_fns.R")
}
\description{
Write wrappers
}
\concept{ui
Writes convenient wrappers around
\code{column(12, ...)}, \code{column(6, ...)}, \code{column(4, ...)}...
named \code{col_1}, \code{col_2} etc.}
