% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{acc_list}
\alias{acc_list}
\title{Create an accordion from a list of items}
\usage{
acc_list(
  x,
  id = "accordion",
  width = 12,
  collapsed = TRUE,
  collapse_all = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{chr/shiny.tag.list} chr vector or shiny.tag.list to wrap in accordions}

\item{id}{Accordion to target.}

\item{width}{The width of the accordion.}

\item{collapsed}{If TRUE, start collapsed. This must be used with
\code{collapsible=TRUE}.}

\item{collapse_all}{\code{lgl} Whether to keep all accordions collapsed except for the open one, or keep them open after opening.}

\item{...}{
  Arguments passed on to \code{\link[bs4Dash:accordion]{bs4Dash::accordion}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\code{shiny.tag.list}
}
\description{
Create an accordion from a list of items
}
\examples{
acc_list(shiny::tagList(shiny::p("a"), shiny::p("b")))
# purrr::map is used internally so the first layer of list nesting determines what resides in each accordion
acc_list(list("a" , shiny::tagList(shiny::p("a"), shiny::p("b"))))
}
\seealso{
Other ui: 
\code{\link{add_sass}()},
\code{\link{box_list}()},
\code{\link{bs4Alert}()},
\code{\link{col_auto}()},
\code{\link{css_props}()},
\code{\link{dynamic_row}()},
\code{\link{enurl}()},
\code{\link{jq_hide}()},
\code{\link{list_to_li}()},
\code{\link{list_to_p}()},
\code{\link{make_action_button}()},
\code{\link{make_columns}()},
\code{\link{rep_br}()},
\code{\link{tagRemoveAttributes}()},
\code{\link{ui_row}()},
\code{\link{ui_tabs}()},
\code{\link{undisplay}()},
\code{\link{with_red_star}()}
}
\concept{ui}
