% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_js.R
\name{actionButton_toggle_style}
\alias{actionButton_toggle_style}
\title{Add a "Toggle" style to an actionButton}
\usage{
actionButton_toggle_style(btn)
}
\arguments{
\item{btn}{\code{shiny.tag} The actionButton to style}
}
\value{
\code{shiny.tag.list} With the button
}
\description{
Wrap the actionbuttion in this function and once clicked, the button will appear to be in a "pressed" state until the next click.
Useful for toggling a floating panel which is hidden/shown on button click
}
\examples{
library(shiny)
devtools::load_all(pkgload::pkg_path())
ui <- fluidPage(
  shinyVirga::use_shinyVirga(),
  actionButton_toggle_style(
    shiny::actionButton(
      'modal',
      "Open Floating Panel",
      icon = shiny::icon("redo")
    )
  )
)


# Define server logic required to draw a histogram
server <- function(input, output) {

}
shinyApp(ui = ui, server = server)
}
