% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{add_module}
\alias{add_module}
\title{Add a module file to the \code{path} specified}
\usage{
add_module(
  name,
  path = "R",
  export = FALSE,
  ph_ui = " ",
  ph_server = " ",
  use_ud = TRUE,
  ...
)
}
\arguments{
\item{name}{\code{chr} name of the module}

\item{path}{\code{chr} directory in which to create it}

\item{export}{\code{lgl} whether it should be exported in the package}

\item{use_ud}{\code{lgl} Whether to include the \code{session$userData} environment as a shorthand variable \code{ud} in the module for ease of access to a user's session state variables.}
}
\value{
\code{file} The new file opens
}
\description{
Add a module file to the \code{path} specified
}
