% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{add_pseudo_module}
\alias{add_pseudo_module}
\title{Add a pseudo-module, one which maintains the namespace of it's parent module}
\usage{
add_pseudo_module(
  name,
  path = "R",
  export = FALSE,
  ph_name = sprintf("pmod_\%s", name),
  ph_ui = sprintf("pmod_\%s_ui", name),
  ph_server = sprintf("pmod_\%s_server", name),
  open = TRUE
)
}
\arguments{
\item{name}{\code{chr} name of the module}

\item{path}{\code{chr} directory in which to create it}

\item{export}{\code{lgl} whether it should be exported in the package}

\item{ph_name}{\code{chr} The name of file}

\item{ph_ui}{\code{chr} The name of the module ui function}

\item{ph_server}{\code{chr} The name of the module server function}
}
\value{
\code{file} The new file opens
}
\description{
Add a pseudo-module, one which maintains the namespace of it's parent module
}
