% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{copyright}
\alias{copyright}
\title{Creates an always up to date italicized copyright element.}
\usage{
copyright(
  ...,
  copyright_holder = shiny::a(logo, href = "https://www.virgalabs.io/", target =
    "_blank"),
  logo = img_base64(css_props = list(`max-width` = "3em", `max-height` = "1em", display =
    "inline-block;")),
  class = ""
)
}
\arguments{
\item{copyright_holder}{\code{shiny.tag} to display for the copyright}

\item{logo}{\code{shiny.tag} A logo to display after the copyright message. Set to NULL to turn off.}

\item{class}{\code{chr} additional classes for the parent div}
}
\value{
\code{shiny.tag}
}
\description{
Creates an always up to date italicized copyright element.
}
\seealso{
Other ui: 
\code{\link{acc_list}()},
\code{\link{add_sass}()},
\code{\link{box_list}()},
\code{\link{bs4Alert}()},
\code{\link{bs_extract_status}()},
\code{\link{bs_statuses}},
\code{\link{col_10}()},
\code{\link{col_11}()},
\code{\link{col_12}()},
\code{\link{col_1}()},
\code{\link{col_2}()},
\code{\link{col_3}()},
\code{\link{col_4}()},
\code{\link{col_5}()},
\code{\link{col_6}()},
\code{\link{col_7}()},
\code{\link{col_8}()},
\code{\link{col_9}()},
\code{\link{col_auto}()},
\code{\link{css_props}()},
\code{\link{dynamic_row}()},
\code{\link{enurl}()},
\code{\link{fa_arrow_icon}()},
\code{\link{icon_sb}()},
\code{\link{infoIcon}()},
\code{\link{jq_hide}()},
\code{\link{list_to_li}()},
\code{\link{list_to_p}()},
\code{\link{make_action_button}()},
\code{\link{make_columns}()},
\code{\link{rep_br}()},
\code{\link{tagRemoveAttributes}()},
\code{\link{ui_row}()},
\code{\link{ui_tabs}()},
\code{\link{undisplay}()},
\code{\link{with_red_star}()}
}
\concept{ui}
