% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{dbg_msg}
\alias{dbg_msg}
\title{Print a debug message to the R and Javascript Console}
\usage{
dbg_msg(x, e = rlang::caller_env())
}
\arguments{
\item{x}{\code{chr} message formatted for \link[cli]{format_inline}}

\item{e}{\code{env} Environment}
}
\value{
\code{msg}
}
\description{
Print a debug message to the R and Javascript Console
}
