% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_deploy.R
\name{deploy_stage}
\alias{deploy_stage}
\title{Run Deployment as a background task}
\usage{
deploy_stage(
  deploy_path = "deploy",
  docker_image_tag = tolower(golem::get_golem_name()),
  use_renv = TRUE,
  copy_r_environ = TRUE,
  copy_r_profile = TRUE,
  copy_renv_lock = TRUE,
  copy_files = NULL,
  rebuild = TRUE,
  remove_previous_builds = TRUE,
  launch_job = TRUE,
  lockfile_path = "renv.lock",
  GITHUB_PAT = remotes:::github_pat()
)
}
\arguments{
\item{deploy_path}{\code{chr} directory in which to stage deployment files}

\item{docker_image_tag}{\code{chr} Docker Image tag name \strong{Default: \code{tolower(golem::get_golem_name())}}}

\item{use_renv}{\code{lgl} whether to use the renv.lock & \code{golem::add_dockerfile_with_renv} (\code{TRUE}) or the DESCRIPTION \code{golem::add_dockerfile}}

\item{copy_r_environ}{\code{lgl} Whether to include the \emph{.Renviron} file}

\item{copy_r_profile}{\code{lgl} Whether to include the \emph{.Rprofile} file}

\item{copy_renv_lock}{\code{lgl} Whether to include the \emph{renv.lock} file}

\item{copy_files}{\code{chr} vector of files to copy in addition to default.}

\item{rebuild}{\code{lgl} Whether to rebuild the tar.gz}

\item{remove_previous_builds}{\code{lgl} Whether to remove previous build files from \code{deploy_path} folder}

\item{launch_job}{\code{lgl} whether to launch the deployment script \emph{jobs.R} in the \code{deploy_path} directory after writing it.}

\item{GITHUB_PAT}{\code{chr} Github Personal Access Token. ** Default remotes:::github_pat() **. Includes the Github Personal Access Token in \emph{.Renviron} copied to Docker image for fetching Github repos during package installation. Set to \code{NULL} to turn off.}
}
\value{
\code{job} Background job that deploys the dmdu app (not working locally)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other deploy: 
\code{\link{deploy_tar}()}
}
\concept{deploy}
