% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_icons.R
\name{icon_sb}
\alias{icon_sb}
\title{Wrapper for \link[shiny]{icon} that supports svg images}
\usage{
icon_sb(name, class = NULL, lib = "font-awesome", ...)
}
\arguments{
\item{name}{\code{chr} FontAwesome name or image path}

\item{class}{Additional classes to customize the style of an icon (see the
\href{https://fontawesome.com/how-to-use}{usage examples} for details on
supported styles).}

\item{lib}{The icon library to use. Either \code{"font-awesome"} or \code{"glyphicon"}.}

\item{...}{Arguments passed to the \verb{<i>} tag of \link[htmltools:builder]{htmltools::tags}.}
}
\value{
An \verb{<i>} (icon) HTML tag.
}
\description{
Wrapper for \link[shiny]{icon} that supports svg images
}
\examples{
# add an icon to a submit button
submitButton("Update View", icon = icon("redo"))

navbarPage("App Title",
  tabPanel("Plot", icon = icon("bar-chart-o")),
  tabPanel("Summary", icon = icon("list-alt")),
  tabPanel("Table", icon = icon("table"))
)
}
