% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_js.R
\name{js_after}
\alias{js_after}
\title{Add a temporary message after an element
Message is added directly below element and persists for three seconds}
\usage{
js_after(
  id,
  content,
  status = NULL,
  properties = list(`font-size` = "12px", display = "block", position = "relative"),
  delay = 3000,
  asis = FALSE,
  .ns = ns_find()
)
}
\arguments{
\item{id}{\code{(chr)} Automatically namespaced with \code{ns} if inside a module}

\item{content}{\code{chr/shiny.tag/shiny.tag.list} Content to add beneath the element}

\item{properties}{\code{list} of css rules to apply to the parent element of the content. Defaults to jQuery validation styled messages.}

\item{delay}{\code{dbl} duration in ms before removing the content from the element. Set to 0 to allow the element to persist.}

\item{asis}{If \code{TRUE}, use the ID as-is even when inside a module
(instead of adding the namespace prefix to the ID).}

\item{.ns}{\code{fun} ns function. Typically found automatically.}
}
\value{
Content added to element
}
\description{
Add a temporary message after an element
Message is added directly below element and persists for three seconds
}
\seealso{
Other JS: 
\code{\link{js_callback}()},
\code{\link{js_callout}()},
\code{\link{js_glow}()},
\code{\link{js_mouseover_once}()},
\code{\link{js_set_input_val}()},
\code{\link{use_driver.js}()}
}
\concept{JS}
