% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_js.R
\name{js_callout}
\alias{js_callout}
\title{Create a driver.js callout
Must include \code{shinyVirga::use_driver.js()} & \code{\link[shinyjs]{useShinyJS}} in the head of the page.}
\usage{
js_callout(
  el,
  title,
  description,
  side = NULL,
  align = NULL,
  animate = TRUE,
  overlayColor = "black",
  smoothScroll = FALSE,
  allowClose = TRUE,
  overlayOpacity = 0.5,
  stagePadding = 10,
  stageRadius = 5,
  allowKeyboardControl = TRUE,
  disableActiveInteraction = FALSE,
  popoverClass = "scoped-class",
  popoverOffset = 10,
  showButtons = NULL,
  disableButtons = NULL,
  showProgress = FALSE,
  progressText = FALSE,
  nextBtnText = "Next",
  prevBtnText = "Previous",
  doneBtnText = "Done",
  onPopoverRender = NULL,
  onHighlightStarted = NULL,
  onHighlighted = NULL,
  onDeselected = NULL,
  onDestroyStarted = NULL,
  onDestroyed = NULL,
  onNextClick = NULL,
  onPrevClick = NULL,
  onCloseClick = NULL,
  asis = FALSE,
  .ns = ns_find()
)
}
\arguments{
\item{el}{\code{chr} ID or class of element. If a character not prefixed by \code{.} for a class or \verb{#} for an ID, it will be assumed and ID and \verb{#} will be added.}

\item{title}{\code{chr} Title of popover}

\item{description}{\code{chr} Description for it}

\item{side}{\code{chr}  The position of the popover relative to the target element. One of "top", "right", "bottom", "left".}

\item{align}{\code{chr} The alignment of the popover relative to the target element. One of "start", 'center', "end".}

\item{animate}{\code{lgl} Whether to animate the product tour. (default: true)}

\item{overlayColor}{\code{chr} Overlay color. (default: black) This is useful when you have a dark background and want to highlight elements with a light background color.}

\item{smoothScroll}{\code{lgl}  Whether to smooth scroll to the highlighted element. (default: false)}

\item{allowClose}{\code{lgl}  Whether to allow closing the popover by clicking on the backdrop. (default: true)}

\item{overlayOpacity}{\code{num}  Opacity of the backdrop. (default: 0.5)}

\item{stagePadding}{\code{num} Distance between the highlighted element and the cutout. (default: 10)}

\item{stageRadius}{\code{num}  Radius of the cutout around the highlighted element. (default: 5)}

\item{allowKeyboardControl}{\code{lgl} Whether to allow keyboard navigation. (default: true)}

\item{disableActiveInteraction}{\code{lgl} Whether to disable interaction with the highlighted element. (default: false)}

\item{popoverClass}{\code{chr} If you want to add custom class to the popover}

\item{popoverOffset}{\code{num} Distance between the popover and the highlighted element. (default: 10)}

\item{showButtons}{\code{chr}  Array of buttons to show in the popover. Defaults to \link{"next", "previous", "close"} for product tours and [] for single element highlighting.}

\item{disableButtons}{\code{chr}  Array of buttons to disable. This is useful when you want to show some of the buttons, but disable some of them.}

\item{showProgress}{\code{lgl}  Whether to show the progress text in popover. (default: false)}

\item{progressText}{\code{chr} See \code{\href{https://driverjs.com/docs/configuration}{docs}} for details}

\item{nextBtnText}{\code{chr} See \code{\href{https://driverjs.com/docs/configuration}{docs}} for details}

\item{prevBtnText}{\code{chr} See \code{\href{https://driverjs.com/docs/configuration}{docs}} for details}

\item{doneBtnText}{\code{chr} See \code{\href{https://driverjs.com/docs/configuration}{docs}} for details}

\item{onPopoverRender}{\code{chr} See \code{\href{https://driverjs.com/docs/configuration}{docs}} for details}

\item{onHighlightStarted}{\code{chr} See \code{\href{https://driverjs.com/docs/configuration}{docs}} for details}

\item{onHighlighted}{\code{chr} See \code{\href{https://driverjs.com/docs/configuration}{docs}} for details}

\item{onDeselected}{\code{chr} See \code{\href{https://driverjs.com/docs/configuration}{docs}} for details}

\item{onDestroyStarted}{\code{chr} See \code{\href{https://driverjs.com/docs/configuration}{docs}} for details}

\item{onDestroyed}{\code{chr} See \code{\href{https://driverjs.com/docs/configuration}{docs}} for details}

\item{onNextClick}{\code{chr} See \code{\href{https://driverjs.com/docs/configuration}{docs}} for details}

\item{onPrevClick}{\code{chr} See \code{\href{https://driverjs.com/docs/configuration}{docs}} for details}

\item{onCloseClick}{\code{chr} See \code{\href{https://driverjs.com/docs/configuration}{docs}} for details}

\item{asis}{If \code{TRUE}, use the ID as-is even when inside a module
(instead of adding the namespace prefix to the ID).}

\item{.ns}{\code{fun} ns function. Typically found automatically.}

\item{closeBtnText}{\code{chr} See \code{\href{https://driverjs.com/docs/configuration}{docs}} for details}
}
\description{
Create a driver.js callout
Must include \code{shinyVirga::use_driver.js()} & \code{\link[shinyjs]{useShinyJS}} in the head of the page.
}
\examples{

js_callout('my_id', "Tooltip title", "Tooltip description")
}
\seealso{
Other JS: 
\code{\link{js_after}()},
\code{\link{js_callback}()},
\code{\link{js_glow}()},
\code{\link{js_mouseover_once}()},
\code{\link{js_picker_disable}()},
\code{\link{js_picker_enable}()},
\code{\link{js_set_input_val}()},
\code{\link{use_driver.js}()}
}
\concept{JS}
