% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_js.R
\name{js_glow}
\alias{js_glow}
\title{Add an animated glow to an element}
\usage{
js_glow(id, color = "deepskyblue", asis = FALSE, delay = NULL, .ns = ns_find())
}
\arguments{
\item{id}{\code{(chr)} Automatically namespaced with \code{ns} if inside a module}

\item{color}{\code{chr} A hexadecimal color to animate with}

\item{asis}{If \code{TRUE}, use the ID as-is even when inside a module
(instead of adding the namespace prefix to the ID).}

\item{delay}{\code{dbl} duration in ms before removing the content from the element. Set to 0 to allow the element to persist.}

\item{.ns}{\code{fun} ns function. Typically found automatically.}
}
\description{
Add an animated glow to an element
}
\examples{
\dontrun{
library(shiny)
devtools::load_all(pkgload::pkg_path())
# Define UI for application that draws a histogram
ui <- fluidPage(
  shinyjs::useShinyjs(),
  div(id = "test", "blah", style = "width:50px"),
)


# Define server logic required to draw a histogram
server <- function(input, output) {

  shinyVirga::js_glow(
    id = "test"
  )

}
}

shinyApp(ui = ui, server = server)
}
\seealso{
Other JS: 
\code{\link{js_after}()},
\code{\link{js_callback}()},
\code{\link{js_callout}()},
\code{\link{js_mouseover_once}()},
\code{\link{js_picker_disable}()},
\code{\link{js_picker_enable}()},
\code{\link{js_set_input_val}()},
\code{\link{use_driver.js}()}
}
\concept{JS}
