% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_golem.R
\name{list_to_p}
\alias{list_to_p}
\title{Turn an R list into corresponding HTML paragraph tags}
\usage{
list_to_p(list, class = NULL)
}
\arguments{
\item{list}{an R list}

\item{class}{a class for the paragraph tags}
}
\value{
An HTML tag
}
\description{
Turn an R list into corresponding HTML paragraph tags
}
\examples{
list_to_p(c("This is the first paragraph", "this is the second paragraph"))

}
\seealso{
Other golem: 
\code{\link{enurl}()},
\code{\link{is_shiny.tag}()},
\code{\link{jq_hide}()},
\code{\link{list_to_li}()},
\code{\link{make_action_button}()},
\code{\link{rep_br}()},
\code{\link{tagRemoveAttributes}()},
\code{\link{undisplay}()},
\code{\link{with_red_star}()}

Other ui: 
\code{\link{acc_list}()},
\code{\link{add_sass}()},
\code{\link{box_list}()},
\code{\link{bs4Alert}()},
\code{\link{col_auto}()},
\code{\link{css_props}()},
\code{\link{dynamic_row}()},
\code{\link{enurl}()},
\code{\link{jq_hide}()},
\code{\link{list_to_li}()},
\code{\link{make_action_button}()},
\code{\link{make_columns}()},
\code{\link{rep_br}()},
\code{\link{tagRemoveAttributes}()},
\code{\link{ui_row}()},
\code{\link{ui_tabs}()},
\code{\link{undisplay}()},
\code{\link{with_red_star}()}
}
\concept{golem}
\concept{ui}
