% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_golem.R
\name{make_action_button}
\alias{make_action_button}
\title{Make the current tag behave like an action button}
\usage{
make_action_button(tag, inputId = NULL)
}
\arguments{
\item{tag}{Any compatible tag.}

\item{inputId}{Unique id. This will host the input value to be used
on the server side.}
}
\value{
The modified tag with an extra id and the action button class.
}
\description{
Only works with compatible tags like button or links
}
\examples{
if (interactive()) {
 library(shiny)

 link <- a(href = "#", "My super link", style = "color: lightblue;")

 ui <- fluidPage(
  make_action_button(link, inputId = "mylink")
 )

 server <- function(input, output, session) {
   observeEvent(input$mylink, {
    showNotification("Pouic!")
   })
 }

 shinyApp(ui, server)

}
}
\seealso{
Other golem: 
\code{\link{enurl}()},
\code{\link{is_shiny.tag}()},
\code{\link{jq_hide}()},
\code{\link{list_to_li}()},
\code{\link{list_to_p}()},
\code{\link{rep_br}()},
\code{\link{tagRemoveAttributes}()},
\code{\link{undisplay}()},
\code{\link{with_red_star}()}

Other ui: 
\code{\link{acc_list}()},
\code{\link{add_sass}()},
\code{\link{box_list}()},
\code{\link{bs4Alert}()},
\code{\link{col_auto}()},
\code{\link{css_props}()},
\code{\link{dynamic_row}()},
\code{\link{enurl}()},
\code{\link{jq_hide}()},
\code{\link{list_to_li}()},
\code{\link{list_to_p}()},
\code{\link{make_columns}()},
\code{\link{rep_br}()},
\code{\link{tagRemoveAttributes}()},
\code{\link{ui_row}()},
\code{\link{ui_tabs}()},
\code{\link{undisplay}()},
\code{\link{with_red_star}()}
}
\concept{golem}
\concept{ui}
