% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{ns_find}
\alias{ns_find}
\title{Retrieve the \code{ns} function}
\usage{
ns_find(e = rlang::caller_env(2))
}
\arguments{
\item{e}{\code{env}}
}
\value{
\code{fun}
}
\description{
Designed to find the \code{ns} function from any level of nesting within a UI or Server module. \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other general: 
\code{\link{browser_server}()},
\code{\link{browser_ui}()},
\code{\link{is_shiny.tag}()},
\code{\link{ns_custom}()},
\code{\link{tab_ns_extract}()},
\code{\link{warn_id}()}
}
\concept{general}
