% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_js.R
\name{render_ran_input}
\alias{render_ran_input}
\title{Create an incrementing shiny input each time a render function runs}
\usage{
render_ran_input(outputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{\code{chr} The outputId value}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}. Default is \code{getDefaultReactiveDomain()}.}
}
\value{
Called for side effects. Runs code via shinyjs.
}
\description{
Requires \code{\link[shinyjs]{runjs}}
}
\examples{
\dontrun{
output[[outputId]] <- renderDT({
  ## DT CODE
  render_ran_input(outputId)
})
}

}
