% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reactives.R
\name{rv_modify}
\alias{rv_modify}
\title{Add named objects to a \code{reactiveValues} object}
\usage{
rv_modify(x, ...)
}
\arguments{
\item{x}{\code{reactiveValues}}

\item{...}{\code{named objects} to assign when \code{assign = TRUE}}
}
\value{
\code{none} Modifies in place
}
\description{
Add named objects to a \code{reactiveValues} object
}
