% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reactives.R
\name{rv_modify}
\alias{rv_modify}
\title{Add named objects to a \code{reactiveValues} object}
\usage{
rv_modify(x, ..., recursive = FALSE, modify_in_place = TRUE)
}
\arguments{
\item{x}{\code{reactiveValues}}

\item{...}{\code{named objects} to assign when \code{assign = TRUE}}

\item{recursive}{\code{lgl} recursively assigns variables into a nested reactiveValues. If \code{FALSE}, the default, values are assigned at the first level of the reactiveValues object.}

\item{modify_in_place}{\code{lgl} modifies the reactiveValues object in place, making it such that assignment is not necessary. If \code{FALSE}, the default, a list copy of the reactiveValues are made and returned. Useful for checking expected output.}
}
\value{
\code{none} Modifies in place, unless \code{modify_in_place = FALSE}
}
\description{
Add named objects to a \code{reactiveValues} object
}
